erp/requirements/base.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==23.2.0
CacheControl==0.14.1
cachetools==5.5.0
certifi==2025.4.26
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cryptography==42.0.5
cssselect2==0.7.0
dateutils==0.6.12
Django==5.0.4
django-binary-database-files==1.0.18
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
django-oauth-toolkit==2.3.0
django-opensearch-dsl==0.7.0
django-vite @ https://github.com/pandosima/django-vite/archive/refs/heads/pandosima/main.zip
djangorestframework==3.15.1
djangorestframework-api-key==2.3.0
docopt==0.6.2
drf-nested-forms==1.1.8
drf-nested-routers==0.93.5
elastic-transport==8.17.1
Events==0.5
firebase-admin==6.6.0
flatbuffers==24.3.25
frozenlist==1.4.1
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
html5lib==1.1
httplib2==0.22.0
humanfriendly==10.0
idna==3.7
inflect==7.3.1
joblib==1.4.2
jwcrypto==1.5.6
lxml==5.3.0
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
mysqlclient==2.2.7
num2words==0.5.14
numpy==1.26.4
oauthlib==3.2.2
onnx==1.16.0
onnxconverter-common @ https://github.com/Skywoodthefifth/onnxconverter-common/archive/refs/heads/master.zip#sha256=5009802c93af0164de0f25e710128993db0731505f212631504e9d768abec4f7
onnxruntime==1.17.3
opensearch-py==2.8.0
oscrypto==1.3.0
packaging==24.0
pandas==2.2.2
phonenumbers==8.13.40
pillow==10.3.0
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
PyMySQL==1.1.1
pyparsing==3.2.0
pypdf==4.3.1
pypng==0.20220715.0
pyreadline3==3.4.1
python-bidi==0.6.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.4
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
skl2onnx==1.16.0
sqlparse==0.4.4
svglib==1.5.1
sympy==1.12
threadpoolctl==3.5.0
tinycss2==1.3.0
twilio==9.2.3
typeguard==4.3.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.4.0
webencodings==0.5.1
xhtml2pdf==0.2.16
yarl==1.9.4
